web-crawlers/brawl-stars/pyproject.toml
[tool.poetry]
name = "web-crawler-brawl-stars"
version = "0.0.1"
description = "Brawl Stars API web crawler."
authors = ["Anthony Caliani <avcaliani.it@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
google-cloud-storage = "^2.7.0"

[tool.poetry.dev-dependencies]
Flask = "^2.2.2"
autopep8 = "^2.0.1"
functions-framework = "^3.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
